+++ src/version/index.js
@@ -12,9 +12,9 @@
 
 // Version info is injected at build time
 // DO NOT EDIT - this is auto-generated
-export const VERSION = '15.0.0';
+export const VERSION = '15.0.1';
 export const BUILD_DATE = '2026-01-24';
-export const BUILD_TIMESTAMP = 1769245953615;
+export const BUILD_TIMESTAMP = 1769250120088;
 
 /**
  * Version information object